rename: 타입선언 위치 및 변수명 수정
