+++ src/types/common.ts
@@ -9,3 +9,24 @@ export interface ApiResponse<T> {
 export interface CustomJwtPayload extends JwtPayload {
   memberId?: string | number;
 }
+
+// 하루미션, 공유일기
+export type WriterInfo = {
+  memberId: number;
+  nickname: string;
+  profileImageUrl: string | null;
+  isMe: boolean;
+};
+
+export type DailyComment = {
+  commentId: number;
+  writerInfo: WriterInfo;
+  content: string;
+  createdAt: string;
+};
+
+export type DailyCommentList = {
+  comments: DailyComment[];
+  lastId: number;
+  hasNext: boolean;
+};
