+++ src/types/diary.type.ts
@@ -1,11 +1,6 @@
+import { WriterInfo } from "./common";
 import { Emotion } from "./emotion.type";
 
-export type WriterInfo = {
-  memberId: number;
-  nickname: string;
-  profileImageUrl: string | null;
-  isMe: boolean;
-};
 // 다이어리 상세 정보 (API 응답과 1:1 매칭)
 export type DiaryDetail = {
   writerInfo: WriterInfo;
@@ -24,19 +19,6 @@ export type DiaryDetail = {
 
 export type DiaryListResponse = {
   diaries: DiaryDetail[];
-  lastId: number;
-  hasNext: boolean;
-};
-
-export type DiaryComment = {
-  commentId: number;
-  writerInfo: WriterInfo;
-  content: string;
-  createdAt: string;
-};
-
-export type DiaryCommentList = {
-  comments: DiaryComment[];
   lastId: number;
   hasNext: boolean;
 };
